requirements.txt
python-dotenv
selenium
turso-python
python-dotenv
